finish demo apiary

+++ src/app/dashboard/apiary/ruche-rucher/ruche.rucher.component.ts
@@ -17,6 +17,8 @@ import { RucherModel } from '../../../_model/rucher-model';
 import { AuthService } from '../../../auth/Service/auth.service';
 import { RucheInterface } from '../../../_model/ruche';
 import { NotifierService } from 'angular-notifier';
+import { MyNotifierService } from '../../service/my-notifier.service';
+import { NotifList } from '../../../../constants/notify';
 
 @Component({
   selector: 'app-ruche-rucher',
@@ -51,6 +53,7 @@ export class RucheRucherComponent implements OnInit, OnDestroy {
 
   constructor(private formBuilder: FormBuilder,
     public location: Location,
+    private myNotifer: MyNotifierService,
     public router: Router,
     public rucherService: RucherService,
     public userService: UserloggedService,
@@ -100,39 +103,47 @@ export class RucheRucherComponent implements OnInit, OnDestroy {
 
   //Pour effacer une ruche
   deleteRuche(ruche: RucheInterface, index: number) {
-    this.rucheService.deleteRuche(index, ruche).subscribe(() => { }, () => { }, () => {
-      this.rucheService.ruches.splice(index, 1);
-      this.rucheService.emitHiveSubject();
-      if(this.userService.getJwtReponse().country === "FR"){
-        this.notify.notify('success', 'Ruche supprimée');
-      }else{
-        this.notify.notify('success', 'Deleted Hive');
-      }
-    });
+    if (this.rucherService.rucher.idUsername === this.userService.getIdUserLoged()) {
+      this.rucheService.deleteRuche(index, ruche).subscribe(() => { }, () => { }, () => {
+        this.rucheService.ruches.splice(index, 1);
+        this.rucheService.emitHiveSubject();
+        if(this.userService.getJwtReponse().country === "FR"){
+          this.notify.notify('success', 'Ruche supprimée');
+        }else{
+          this.notify.notify('success', 'Deleted Hive');
+        }
+      });
+    } else {
+      this.myNotifer.sendWarningNotif(NotifList.AUTH_WRITE_HIVE);
+    }
   }
 
   //Pour créer une ruche
   createRuche() {
-    const formValue = this.newRucheForm.value;
-    this.selectHive.id = null;
-    this.selectHive.idApiary = this.rucherService.rucher.id;
-    this.selectHive.idUsername = this.userService.getIdUserLoged();
-    this.selectHive.description = formValue.descriptionRuche;
-    this.selectHive.name = formValue.nomRuche;
-    this.selectHive.apiaryName = this.rucherService.rucher.name;
-    this.selectHive.username = this.username.toLowerCase();
-    this.initForm();
-    this.rucheService.createRuche(this.selectHive).subscribe((hive) => {
-      // this.rucheService.saveCurrentHive(hive.id);
-      this.rucheService.ruches.push(hive);
-    }, () => { }, () => {
-      this.rucheService.emitHiveSubject();
-      if(this.userService.getJwtReponse().country === "FR"){
-        this.notify.notify('success', 'Ruche créée');
-      }else{
-        this.notify.notify('success', 'Crated Hive');
-      }
-    });
+    if (this.rucherService.rucher.idUsername === this.userService.getIdUserLoged()) {
+      const formValue = this.newRucheForm.value;
+      this.selectHive.id = null;
+      this.selectHive.idApiary = this.rucherService.rucher.id;
+      this.selectHive.idUsername = this.userService.getIdUserLoged();
+      this.selectHive.description = formValue.descriptionRuche;
+      this.selectHive.name = formValue.nomRuche;
+      this.selectHive.apiaryName = this.rucherService.rucher.name;
+      this.selectHive.username = this.username.toLowerCase();
+      this.initForm();
+      this.rucheService.createRuche(this.selectHive).subscribe((hive) => {
+        // this.rucheService.saveCurrentHive(hive.id);
+        this.rucheService.ruches.push(hive);
+      }, () => { }, () => {
+        this.rucheService.emitHiveSubject();
+        if(this.userService.getJwtReponse().country === "FR"){
+          this.notify.notify('success', 'Ruche créée');
+        }else{
+          this.notify.notify('success', 'Crated Hive');
+        }
+      });
+    } else {
+      this.myNotifer.sendWarningNotif(NotifList.AUTH_WRITE_HIVE);
+    }
   }
 
   onSelectRuche(ruche: RucheInterface, index: number) {
@@ -147,21 +158,25 @@ export class RucheRucherComponent implements OnInit, OnDestroy {
   }
   // pour editer une ruche
   onEditeRuche() {
-    const formValue = this.newRucheForm.value;
-    this.selectHive.idApiary = this.rucherService.rucherSelectUpdate.id;
-    this.selectHive.name = formValue.nomRuche;
-    this.selectHive.description = formValue.descriptionRuche;
-    this.rucheService.updateRuche(this.hiveIndex, this.selectHive).subscribe(() => { }, () => { }, () => {
-      if (this.selectHive.idApiary === this.rucherService.getCurrentApiary()) {
-        this.rucheService.ruches[this.hiveIndex] = this.selectHive;
-        this.rucheService.emitHiveSubject();
-      }
-      if(this.userService.getJwtReponse().country === "FR"){
-        this.notify.notify('success', 'Ruche mis à jour');
-      }else{
-        this.notify.notify('success', 'Updated Hive');
-      }
-    });
+    if (this.rucherService.rucher.idUsername === this.userService.getIdUserLoged()) {
+      const formValue = this.newRucheForm.value;
+      this.selectHive.idApiary = this.rucherService.rucherSelectUpdate.id;
+      this.selectHive.name = formValue.nomRuche;
+      this.selectHive.description = formValue.descriptionRuche;
+      this.rucheService.updateRuche(this.hiveIndex, this.selectHive).subscribe(() => { }, () => { }, () => {
+        if (this.selectHive.idApiary === this.rucherService.getCurrentApiary()) {
+          this.rucheService.ruches[this.hiveIndex] = this.selectHive;
+          this.rucheService.emitHiveSubject();
+        }
+        if(this.userService.getJwtReponse().country === "FR"){
+          this.notify.notify('success', 'Ruche mis à jour');
+        }else{
+          this.notify.notify('success', 'Updated Hive');
+        }
+      });
+    } else {
+      this.myNotifer.sendWarningNotif(NotifList.AUTH_WRITE_HIVE);
+    }
   }
 
   editRucherClicked() {
